# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Task API",
    "version": "1.0.0",
    "description": "A simple CRUD API for managing a to-do list built during FlyRank Internship Week 2."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/": {
      "get": {
        "summary": "Root endpoint describing the API",
        "responses": {
          "200": { "description": "Success" }
        }
      }
    },
    "/health": {
      "get": {
        "summary": "Health check endpoint",
        "responses": {
          "200": { "description": "Server is alive" }
        }
      }
    },
    "/tasks": {
      "get": {
        "summary": "Get all tasks",
        "responses": {
          "200": { "description": "Success" }
        }
      },
      "post": {
        "summary": "Create a new task",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" }
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request - Missing title" }
        }
      }
    },
    "/tasks/{id}": {
      "get": {
        "summary": "Get a specific task by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": { "description": "Success" },
          "404": { "description": "Task not found" }
        }
      },
      "put": {
        "summary": "Update a task by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "done": { "type": "boolean" }
                },
                "required": ["title"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Success" },
          "400": { "description": "Bad Request" },
          "404": { "description": "Task not found" }
        }
      },
      "delete": {
        "summary": "Delete a task by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": {
          "204": { "description": "No Content" },
          "404": { "description": "Task not found" }
        }
      }
    }
  }
}